+++ src/Equinor.ProCoSys.Completion.WebApi/Controllers/PatchDtoValidator.cs
@@ -1,5 +1,4 @@
-﻿using Equinor.ProCoSys.Completion.WebApi.InputValidators;
-using FluentValidation;
+﻿using FluentValidation;
 using Microsoft.AspNetCore.JsonPatch;
 
 namespace Equinor.ProCoSys.Completion.WebApi.Controllers;
