+++ src/tests/Equinor.ProCoSys.Completion.WebApi.Tests/Controllers/RowVersionDtoValidatorTests.cs
@@ -1,6 +1,5 @@
 ﻿using System.Threading.Tasks;
 using Equinor.ProCoSys.Completion.WebApi.Controllers;
-using Equinor.ProCoSys.Completion.WebApi.InputValidators;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using NSubstitute;
