+++ src/tests/Equinor.ProCoSys.Completion.WebApi.Tests/Controllers/RowVersionValidatorTests.cs
@@ -1,7 +1,7 @@
-﻿using Equinor.ProCoSys.Completion.WebApi.InputValidators;
+﻿using Equinor.ProCoSys.Completion.WebApi.Controllers;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Equinor.ProCoSys.Completion.WebApi.Tests.InputValidators;
+namespace Equinor.ProCoSys.Completion.WebApi.Tests.Controllers;
 
 [TestClass]
 public class RowVersionValidatorTests
