+++ src/tests/Equinor.ProCoSys.Completion.WebApi.Tests/PatchDtoValidatorTests.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using Equinor.ProCoSys.Completion.WebApi.Controllers;
-using Equinor.ProCoSys.Completion.WebApi.InputValidators;
 using Microsoft.AspNetCore.JsonPatch.Operations;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using NSubstitute;
